+++ tests/testthat/test-markdown.R
@@ -591,6 +591,19 @@ yes
   )
 })
 
+test_that("description can start with a heading (#1705)", {
+  text <- "
+    #' Title
+    #'
+    #' # Heading
+    #' @md
+    #' @name x
+    NULL
+  "
+  out <- roc_proc_text(rd_roclet(), text)[[1]]
+  expect_equal(out$get_value("section"), list(title = "Heading", content = ""))
+})
+
 test_that("level >2 markdown headings work in @details", {
   text <- "
     #' Title